+++ environments/production/modules/mcollective_agent_package/files/mcollective/util/package/packagehelpers.rb
@@ -17,10 +17,10 @@ module MCollective
         end
 
         def self.rpm_count(output = "")
-          raise "Cannot find rpm at /usr/bin/rpm" unless File.exists?("/usr/bin/rpm")
+          raise "Cannot find rpm at /bin/rpm" unless File.exists?("/bin/rpm")
           result = {:exitcode => nil,
                     :output => ""}
-          cmd = Shell.new("/usr/bin/rpm -qa", :stdout => output)
+          cmd = Shell.new("/bin/rpm -qa", :stdout => output)
           cmd.runcommand
           result[:exitcode] = cmd.status.exitstatus
           result[:output] = output.split("\n").select{ |line| line != "" }.size.to_s
@@ -55,10 +55,10 @@ module MCollective
 
 
         def self.rpm_md5(output = "")
-          raise "Cannot find rpm at /usr/bin/rpm" unless File.exists?("/usr/bin/rpm")
+          raise "Cannot find rpm at /bin/rpm" unless File.exists?("/bin/rpm")
           result = {:exitcode => nil,
                     :output => ""}
-          cmd = Shell.new("/usr/bin/rpm -qa", :stdout => output)
+          cmd = Shell.new("/bin/rpm -qa", :stdout => output)
           cmd.runcommand
           result[:exitcode] = cmd.status.exitstatus
           result[:output] = Digest::MD5.new.hexdigest(output)
@@ -107,7 +107,10 @@ module MCollective
           result[:exitcode] = cmd.status.exitstatus
 
           raise "apt-get update failed, exit code was #{result[:exitcode]}" unless result[:exitcode] == 0
-          return result
+
+          # Everything was fine.  Discard the current result and return the
+          # actual status of the system.
+          return apt_checkupdates
         end
 
         def self.packagemanager