+++ environments/production/modules/mcollective_choria/files/mcollective/application/playbook.rb
@@ -95,9 +95,9 @@ module MCollective
 
           self.class.option :__loglevel,
                             :arguments => ["--loglevel LEVEL"],
-                            :description => "Override the loglevel set in the playbook (debug, info, warn, error, fatal)",
+                            :description => "Override the loglevel set in the playbook (emerg alert crit err warning notice info verbose debug)",
                             :type => String,
-                            :validate => ->(level) { ["error", "fatal", "debug", "warn", "info"].include?(level) }
+                            :validate => ->(level) { %w[emerg alert crit err warning notice info verbose debug].include?(level) }
 
           super
         end
@@ -184,7 +184,11 @@ module MCollective
           puts
           puts "Result: "
           puts
-          puts Util.align_text(JSON.pretty_generate(result), 10000)
+          if result.class != String
+            puts Util.align_text(JSON.pretty_generate(result), 10000)
+          else
+            puts result
+          end
           puts
         end