+++ environments/production/modules/package/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org).
 
+## [v0.7.0](https://github.com/puppetlabs/puppetlabs-package/tree/v0.7.0) (2019-10-28)
+
+[Full Changelog](https://github.com/puppetlabs/puppetlabs-package/compare/v0.6.0...v0.7.0)
+
+### Added
+
+- \(BOLT-1525\) Zypper implementation [\#143](https://github.com/puppetlabs/puppetlabs-package/pull/143) ([m0dular](https://github.com/m0dular))
+- \(FM-8225\) Convert to litmus testing [\#134](https://github.com/puppetlabs/puppetlabs-package/pull/134) ([sheenaajay](https://github.com/sheenaajay))
+
 ## [v0.6.0](https://github.com/puppetlabs/puppetlabs-package/tree/v0.6.0) (2019-06-12)
 
 [Full Changelog](https://github.com/puppetlabs/puppetlabs-package/compare/v0.5.0...v0.6.0)